# api/handlers/v1/about_h/post.go
package aboutsh

import (
	"context"
	"net/http"

	"github.com/gin-gonic/gin"
	errorh "gitlab.com/azizbekdev-blog/go-monolithic/api/handlers/v1/error_h"
	"gitlab.com/azizbekdev-blog/go-monolithic/storage/postgres/aboutrepo"
)

// @Summary		Create about
// @Tags        About
// @Description	Here about can be created.
// @Security    BearerAuth
// @Accept      json
// @Produce		json
// @Param       post   body      CreateReq true "post info"
// @Success		200 	{object}  models.Response
// @Failure     default {object}  FullResponse
// @Router		/about [POST]
func (h *AboutHandler) Create(c *gin.Context) {
	body := &CreateReq{}
	err := c.ShouldBindJSON(&body)
	if errorh.HandleBadRequestErrWithMessage(c, h.log, err, "c.ShouldBindJSON(&body)") {
		return
	}

	res, err := h.postgres.About().Create(context.Background(), (*aboutrepo.CreateReq)(body))
	if errorh.HandleDatabaseLevelWithMessage(c, h.log, err, "h.postgres.About().Create()") {
		return
	}

	c.JSON(http.StatusOK, &FullResponse{
		ErrorCode:    errorh.ErrorSuccessCode,
		ErrorMessage: "",
		Body:         (*FullResponseBody)(res),
	})
}